analytics/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
redis==5.0.1
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.13.1
